style: Move the serprog logic into a separate file
